chore: Extracted adaptation option into its own class and removed wrong import
